Move planning CLI front door to generated entrypoint
