+++ .agentic-workspace/planning/state.toml
@@ -10,7 +10,7 @@ work_items = []
 
 [active]
 execplans = [
-  { id = "github-892-python-codegen-completion", maturity = "active", status = "active", priority = "P1", refs = "GitHub #892; #930 closed; #931; #933; #946; #947; .agentic-workspace/planning/decompositions/python-generated-cli.decomposition.json; .agentic-workspace/planning/execplans/archive/python-operation-ir-package-resource-commands.plan.json", title = "Finish Python generated CLI and operation-contract completion", outcome = "Complete the Python lane so executable CLI front doors and runtime operation boundaries are generated or represented by implementation-independent artifacts, with no hand-owned cli.py outside generated/python and no implicit Python-internal dependency for future targets.", reason = "The previous completion claim only proved generated adapter coverage; the user clarified that Python is not complete while hand-owned cli.py entrypoints remain and future targets must not silently rely on Python internals.", suggested_first_slice = "Migrate the root agentic-workspace Python console front door away from package-local cli.py while preserving generated dispatch proof.", path = ".agentic-workspace/planning/execplans/python-codegen-completion.plan.json" },
+  { id = "github-892-python-codegen-completion", maturity = "active", status = "active", priority = "P1", refs = "GitHub #892; #930 closed; #931; #933; #946; #947; .agentic-workspace/planning/decompositions/python-generated-cli.decomposition.json; .agentic-workspace/planning/execplans/archive/python-operation-ir-package-resource-commands.plan.json", title = "Finish Python generated CLI and operation-contract completion", outcome = "Complete the Python lane so executable CLI front doors and runtime operation boundaries are generated or represented by implementation-independent artifacts, with no hand-owned cli.py outside generated/python and no implicit Python-internal dependency for future targets.", reason = "The previous completion claim only proved generated adapter coverage; the user clarified that Python is not complete while hand-owned cli.py entrypoints remain and future targets must not silently rely on Python internals.", suggested_first_slice = "Migrate the agentic-memory Python console front door away from package-local cli.py while preserving generated dispatch proof.", path = ".agentic-workspace/planning/execplans/python-codegen-completion.plan.json" },
 ]
 
 [todo]
