+++ packages/planning/pyproject.toml
@@ -29,7 +29,7 @@ Repository = "https://github.com/rickardvh/agentic-workspace"
 Issues = "https://github.com/rickardvh/agentic-workspace/issues"
 
 [project.scripts]
-agentic-planning = "repo_planning_bootstrap.cli:main"
+agentic-planning = "repo_planning_bootstrap.generated_cli_entrypoint:main"
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/repo_planning_bootstrap"]
