+++ scripts/check/measure_lazy_discovery.py
@@ -10,7 +10,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Callable
 
-from repo_planning_bootstrap import cli as planning_cli
+from repo_planning_bootstrap import _runtime_cli as planning_cli
 
 from agentic_workspace import _runtime_cli as workspace_cli
