+++ src/agentic_workspace/_runtime_cli.py
@@ -1621,7 +1621,7 @@ def _memory_module_argv(args: argparse.Namespace) -> list[str]:
 
 
 def _run_planning_front_door(args: argparse.Namespace) -> int:
-    from repo_planning_bootstrap.cli import main as planning_main
+    from repo_planning_bootstrap.generated_cli_entrypoint import main as planning_main
 
     buffer = io.StringIO()
     with contextlib.redirect_stdout(buffer):
@@ -2391,7 +2391,7 @@ def main(argv: list[str] | None = None) -> int:
         try:
             target_root = _resolve_target_root(args.target) if args.target else _resolve_target_root(None)
             _validate_target_root(command_name="summary", target_root=target_root)
-            from repo_planning_bootstrap.cli import _print_summary
+            from repo_planning_bootstrap._runtime_cli import _print_summary
             from repo_planning_bootstrap.installer import format_summary_json, planning_summary
 
             config = _load_workspace_config(target_root=target_root)
@@ -2676,7 +2676,7 @@ def main(argv: list[str] | None = None) -> int:
         try:
             target_root = _resolve_target_root(args.target) if args.target else _resolve_target_root(None)
             _validate_target_root(command_name="reconcile", target_root=target_root)
-            from repo_planning_bootstrap.cli import _print_reconcile
+            from repo_planning_bootstrap._runtime_cli import _print_reconcile
             from repo_planning_bootstrap.installer import planning_reconcile
         except ImportError:
             parser.error("The planning module must be installed to use the reconcile command.")
@@ -19299,7 +19299,7 @@ def _run_start_context_adapter(args: argparse.Namespace) -> int:
 def _run_summary_report_adapter(args: argparse.Namespace) -> int:
     target_root = _resolve_target_root(args.target) if args.target else _resolve_target_root(None)
     _validate_target_root(command_name="summary", target_root=target_root)
-    from repo_planning_bootstrap.cli import _print_summary
+    from repo_planning_bootstrap._runtime_cli import _print_summary
     from repo_planning_bootstrap.installer import format_summary_json, planning_summary
 
     config = _load_workspace_config(target_root=target_root)
@@ -19501,7 +19501,7 @@ def _run_report_combined_adapter(args: argparse.Namespace) -> int:
 def _run_reconcile_report_adapter(args: argparse.Namespace) -> int:
     target_root = _resolve_target_root(args.target) if args.target else _resolve_target_root(None)
     _validate_target_root(command_name="reconcile", target_root=target_root)
-    from repo_planning_bootstrap.cli import _print_reconcile
+    from repo_planning_bootstrap._runtime_cli import _print_reconcile
     from repo_planning_bootstrap.installer import planning_reconcile
 
     _ensure_external_intent_cache_if_available(target_root=target_root)
