+++ src/agentic_workspace/contracts/proof_selection_rules.json
@@ -196,7 +196,7 @@
         "id": "package-cli-runtime",
         "role": "hand-owned-executable",
         "exact": [
-          "packages/planning/src/repo_planning_bootstrap/cli.py",
+          "packages/planning/src/repo_planning_bootstrap/_runtime_cli.py",
           "packages/memory/src/repo_memory_bootstrap/cli.py"
         ],
         "direct_edit_allowed": true,
