+++ tests/test_generated_tool_conformance.py
@@ -26,7 +26,7 @@ CLI_SHIM = (
 PLANNING_CLI_SHIM = (
     "import sys; "
     f"sys.path.insert(0, {str(REPO_ROOT / 'packages' / 'planning' / 'src')!r}); "
-    "from repo_planning_bootstrap.cli import main; "
+    "from repo_planning_bootstrap.generated_cli_entrypoint import main; "
     "raise SystemExit(main(sys.argv[1:]))"
 )
 MEMORY_CLI_SHIM = (
